requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.1
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.6.0
anyio==4.6.2.post1
asyncio==3.4.3
attrs==23.1.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
black==24.10.0
certifi==2023.11.17
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
frozenlist==1.4.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.6
magic-filter==1.0.12
multidict==6.0.4
mypy-extensions==1.0.0
packaging==24.2
paramiko==3.5.0
pathspec==0.12.1
platformdirs==4.3.6
pyaes==1.6.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.5.3
pydantic_core==2.14.6
PyNaCl==1.5.0
python-dotenv==1.0.1
python-telegram-bot==21.6
requests==2.32.3
rsa==4.9
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
Telethon==1.38.1
typing_extensions==4.9.0
urllib3==2.3.0
yarl==1.9.4
